ver1.0/requirements.txt
customtkinter~=5.1.3
tkinter
cryptography~=40.0.2
hashlib

Pillow~=9.5.0